Moved auth middleware to separate file, refactored products routes
