+++ 02-express-tutorial/app.js
@@ -1,9 +1,10 @@
 const express = require('express')
 const app = express()
-const { products, people } = require('./data')
 const peopleRouter = require('./routes/people')
+const productsRouter = require('./routes/products')
 const logger = require('./logger')
 const cookieParser = require('cookie-parser')
+const auth = require('./routes/auth') 
 
 app.use(express.static('./methods-public'))
 app.use(logger)
@@ -11,15 +12,7 @@ app.use(express.urlencoded({ extended: false }))
 app.use(express.json())
 app.use(cookieParser())
 app.use('/api/v1/people', peopleRouter)
-
-const auth = ((req, res, next) => {
-    const name = req.cookies.name
-    if(!name){
-        return res.status(401).json({message: "unauthorized"})
-    }
-    req.user = name
-    next()
-})
+app.use('/api/v1', productsRouter)
 
 app.post('/logon', (req, res)=>{
     const {name} = req.body
@@ -41,43 +34,6 @@ app.get('/test', auth, (req, res)=>{
 
 app.get('/about', (req,res,)=>{
     res.status(200).send("<h1>This is About Page</h1>")
-})
-
-app.get('/api/v1/products', (req, res)=>{ 
-    res.status(200).json(products)
-})
-
-app.get('/api/v1/products/:productID', (req, res)=>{
-    const {productID} = req.params;
-    const singleProduct = products.find((product)=>product.id === Number(productID))
-    if(!singleProduct){
-        return res.status(404).json({message: "That product was not found."})
-    }
-    return res.json(singleProduct)
-})
-
-app.get('/api/v1/query', (req,res)=>{
-    let sortedProducts = [...products]
-    const {search, minPrice, maxPrice, limit} = req.query
-    if(search){
-        sortedProducts = sortedProducts.filter((product)=>{
-            return product.name.startsWith(search)
-        })
-    }
-    if(minPrice || maxPrice){
-        sortedProducts = sortedProducts.filter((product)=>{
-            const meetsMin = minPrice ? product.price >= parseFloat(minPrice) : true;
-            const meetsMax = maxPrice ? product.price <= parseFloat(maxPrice) : true;
-            return  meetsMin && meetsMax
-        })
-    }
-    if(limit){
-        sortedProducts = sortedProducts.slice(0, Number(limit))
-    }
-    if(sortedProducts < 1){
-        res.status(200).send('<h1>No Products Matched Your Search</h1>')
-    }
-    res.status(200).json(sortedProducts)
 })
 
 app.all('*', (req,res)=>{
